--- a/examples/armeria-http4s/src/main/scala/com/exmaple/http4s/armeria/ArmeriaExample.scala
+++ b/examples/armeria-http4s/src/main/scala/com/exmaple/http4s/armeria/ArmeriaExample.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2020-2020 http4s.org
  *
  * SPDX-License-Identifier: Apache-2.0
  */